Rename "space" models, values and primitives to "size"
