+++ docs/designer-decisions/src/content/docs/api/designer-shows/Types/SizeVizName.mdx
@@ -5,7 +5,7 @@ slug: api/designer-shows/Types/SizeVizName
 
 import { ShowDecision } from '@noodlestan/designer-shows/astro';
 
-Available visualizations for [Space Values](/integrations/Astro/Composable/Values/#showsizevalue)
+Available visualizations for [Size Values](/integrations/Astro/Composable/Values/#showsizevalue)
 
 ```ts
 type SizeVizName = 'bar-h' | 'bar-v' | 'square' | 'circle';
