+++ docs/designer-decisions/src/content/docs/integrations/Astro/_partials/TableDecisionComponents.astro
@@ -11,102 +11,125 @@ const types = [
     {
         name: 'SizeValue',
         type: 'size-value',
+        domain: 'space',
         viz: 'ShowSizeViz',
         value: 'ShowSizeValue',
     },
     {
         name: 'SizeScale',
         type: 'size-scale',
+        domain: 'space',
         viz: 'ShowSizeViz',
         value: 'ShowSizeValue',
     },
     {
         name: 'ColorValue',
         type: 'color-value',
+        domain: 'color',
         viz: 'ShowColorViz',
         value: 'ShowColorValue',
     },
     {
         name: 'ColorSet',
         type: 'color-set',
+        domain: 'color',
         viz: 'ShowColorViz',
         value: 'ShowColorValue',
     },
     {
         name: 'Color Oklab Hue Value',
         type: 'color-oklab-hue-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color Oklab Hue Set',
         type: 'color-oklab-hue-set',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color Oklab Chroma Value',
         type: 'color-oklab-chroma-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color Oklab Chroma Scale',
         type: 'color-oklab-chroma-scale',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color Oklab Lightness Value',
         type: 'color-oklab-lightness-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color Oklab Lightness Scale',
         type: 'color-oklab-lightness-scale',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color sRGB Hue Value',
         type: 'color-srgb-hue-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color sRGB Hue Set',
         type: 'color-srgb-hue-set',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color sRGB Saturation Value',
         type: 'color-srgb-saturation-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color sRGB Saturation Scale',
         type: 'color-srgb-saturation-scale',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
         name: 'Color sRGB Lightness Value',
         type: 'color-srgb-lightness-value',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
     {
-        name: 'Color sRGB Lightness Scale ',
-        type: 'color-srgb-lightness-scale ',
+        name: 'Color sRGB Lightness Scale',
+        type: 'color-srgb-lightness-scale',
+        domain: 'color',
         viz: 'ShowColorChannelViz',
         value: 'ShowColorChannelValue',
     },
+    {
+        name: 'Typeface Value ',
+        type: 'typeface-value',
+        domain: 'typography',
+        viz: 'ShowTypeViz',
+        value: 'ShowTypeValue',
+    },
 ];
 
-const items = domain ? types.filter(({ type }) => type.startsWith(domain)) : types;
+const items = domain ? types.filter(({ domain: d }) => d === domain) : types;
 ---
 
 <Table columns={['Decision Type', 'Value Component', 'Viz Component']}>
