+++ packages/libs/designer-decisions/resources/data/samples/space/size-scale-anchored.json
