+++ packages/libs/designer-decisions/resources/data/samples/space/size-value-explicit.json
