+++ packages/libs/designer-decisions/src/inputs/types/models/color/color-value.ts
