+++ packages/libs/designer-decisions/src/inputs/types/models/color/index.ts
@@ -1,5 +1,5 @@
 // @index(['./*.ts', '!./*.test.ts', './!(private|mocks|functions)*/index.ts'], f => `export * from '${f.path.replace(/\/index$/, '')}';`)
-export * from './color';
+export * from './color-set';
+export * from './color-value';
 export * from './oklab';
-export * from './set';
 export * from './srgb';
