+++ packages/libs/designer-decisions/src/inputs/types/models/color/oklab/color-oklab-chroma-value.ts
@@ -1,0 +1,9 @@
+import type { ColorOklabChromaInput, DecisionInput } from '../../../primitives';
+
+export type ColorOklabChromaValueExplicitInput = DecisionInput & {
+    model: 'color-oklab-chroma-value/explicit';
+    params: {
+        value: ColorOklabChromaInput;
+        quantize?: number;
+    };
+};
