+++ packages/libs/designer-decisions/src/inputs/types/models/color/oklab/color-oklab-hue-value.ts
@@ -1,0 +1,9 @@
+import type { ColorOklabHueInput, DecisionInput } from '../../../primitives';
+
+export type ColorOklabHueValueExplicitInput = DecisionInput & {
+    model: 'color-oklab-hue-value/explicit';
+    params: {
+        value: ColorOklabHueInput;
+        quantize?: number;
+    };
+};
