+++ packages/libs/designer-decisions/src/inputs/types/models/color/oklab/color-oklab-lightness-scale.ts
@@ -4,14 +4,6 @@ import type {
     DecisionInput,
 } from '../../../primitives';
 
-export type ColorOklabLightnessValueExplicitInput = DecisionInput & {
-    model: 'color-oklab-lightness-value/explicit';
-    params: {
-        value: ColorOklabLightnessInput;
-        quantize?: number;
-    };
-};
-
 export type ColorOklabLightnessScaleExplicitInput = DecisionInput & {
     model: 'color-oklab-lightness-scale/explicit';
     params: {
