+++ packages/libs/designer-decisions/src/inputs/types/models/color/oklab/color-oklab-lightness-value.ts
@@ -1,0 +1,9 @@
+import type { ColorOklabLightnessInput, DecisionInput } from '../../../primitives';
+
+export type ColorOklabLightnessValueExplicitInput = DecisionInput & {
+    model: 'color-oklab-lightness-value/explicit';
+    params: {
+        value: ColorOklabLightnessInput;
+        quantize?: number;
+    };
+};
