+++ packages/libs/designer-decisions/src/inputs/types/models/color/oklab/index.ts
@@ -1,4 +1,7 @@
 // @index(['./*.ts', '!./*.test.ts', './!(private|mocks|functions)*/index.ts'], f => `export * from '${f.path.replace(/\/index$/, '')}';`)
-export * from './chroma';
-export * from './hue';
-export * from './lightness';
+export * from './color-oklab-chroma-scale';
+export * from './color-oklab-chroma-value';
+export * from './color-oklab-hue-set';
+export * from './color-oklab-hue-value';
+export * from './color-oklab-lightness-scale';
+export * from './color-oklab-lightness-value';
