+++ packages/libs/designer-decisions/src/inputs/types/models/color/srgb/color-srgb-hue-value.ts
@@ -1,0 +1,9 @@
+import type { ColorSRGBHueInput, DecisionInput } from '../../../primitives';
+
+export type ColorSRGBHueValueExplicitInput = DecisionInput & {
+    model: 'color-srgb-hue-value/explicit';
+    params: {
+        value: ColorSRGBHueInput;
+        quantize?: number;
+    };
+};
