+++ packages/libs/designer-decisions/src/inputs/types/models/color/srgb/color-srgb-lightness-value.ts
@@ -1,0 +1,9 @@
+import type { ColorSRGBLightnessInput, DecisionInput } from '../../../primitives';
+
+export type ColorSRGBLightnessValueExplicitInput = DecisionInput & {
+    model: 'color-srgb-lightness-value/explicit';
+    params: {
+        value: ColorSRGBLightnessInput;
+        quantize?: number;
+    };
+};
