+++ packages/libs/designer-decisions/src/inputs/types/models/color/srgb/color-srgb-saturation-value.ts
@@ -1,0 +1,9 @@
+import type { ColorSRGBSaturationInput, DecisionInput } from '../../../primitives';
+
+export type ColorSRGBSaturationValueExplicitInput = DecisionInput & {
+    model: 'color-srgb-saturation-value/explicit';
+    params: {
+        value: ColorSRGBSaturationInput;
+        quantize?: number;
+    };
+};
