+++ packages/libs/designer-decisions/src/inputs/types/models/color/srgb/index.ts
@@ -1,4 +1,7 @@
 // @index(['./*.ts', '!./*.test.ts', './!(private|mocks|functions)*/index.ts'], f => `export * from '${f.path.replace(/\/index$/, '')}';`)
-export * from './hue';
-export * from './lightness';
-export * from './saturation';
+export * from './color-srgb-hue-set';
+export * from './color-srgb-hue-value';
+export * from './color-srgb-lightness-scale';
+export * from './color-srgb-lightness-value';
+export * from './color-srgb-saturation-scale';
+export * from './color-srgb-saturation-value';
