+++ packages/libs/designer-decisions/src/inputs/types/models/space/index.ts
@@ -1,3 +1,3 @@
 // @index(['./*.ts', '!./*.test.ts', './!(private|mocks|functions)*/index.ts'], f => `export * from '${f.path.replace(/\/index$/, '')}';`)
-export * from './set';
-export * from './value';
+export * from './size-scale';
+export * from './size-value';
