+++ packages/libs/designer-decisions/src/inputs/types/models/space/size-value.ts
