+++ packages/libs/designer-decisions/src/inputs/types/models/typography/index.ts
@@ -1,2 +1,2 @@
 // @index(['./*.ts', '!./*.test.ts', './!(private|mocks|functions)*/index.ts'], f => `export * from '${f.path.replace(/\/index$/, '')}';`)
-export * from './value';
+export * from './typeface-value';
