+++ packages/libs/designer-decisions/src/values/primitives/size/helpers/constants.ts
@@ -1,3 +1,3 @@
 import type { SizeAbsoluteUnits } from '../../../../inputs';
 
-export const SIZE_ABSOLUtE_UNITS: SizeAbsoluteUnits[] = ['px', 'rem'];
+export const SIZE_ABSOLUTE_UNITS: SizeAbsoluteUnits[] = ['px', 'rem'];
